+++ src/swagger/client.ts
@@ -35,6 +35,7 @@ import {
   type PortalsListResponse,
   type Product,
   type ProductsListResponse,
+  type PublishPortalProductResponse,
   type PublishProductArgs,
   type ScanApiStandardizationFromRegistryParams,
   type ScanApiStandardizationFromRegistryResult,
@@ -229,9 +230,9 @@ export class SwaggerClient implements Client {
 
   async publishPortalProduct(
     args: PublishProductArgs,
-  ): Promise<SuccessResponse | FallbackResponse> {
-    const { productId, preview } = args;
-    return this.getApi().publishPortalProduct(productId, preview);
+  ): Promise<PublishPortalProductResponse | FallbackResponse> {
+    const { productId, preview, tableOfContentsId } = args;
+    return this.getApi().publishPortalProduct(productId, preview, tableOfContentsId ?? null);
   }
 
   async getPortalProductSections(